+++ changelog.md
@@ -1,3 +1,1 @@
-- Add an api to create a config screen.
-- Prevent changing screen types when it has been forced by the inventory.
-- Prevent opening of null inventories.
+- Fix quilt networking.